backend/requirements.txt
asgiref==3.4.1
Django
gunicorn
psycopg2-binary
sqlparse==0.4.2
tzdata==2021.5
djangorestframework
django-filter
djoser
tqdm
Pillow
django-extra-fields